Second task of Lesson 2 was added
